+++ src/ideas-overview/component/idea-tile.jsx
@@ -14,8 +14,7 @@ export default class IdeasOverview extends OpenStadComponent {
 
   constructor(props) {
 
-		// config
-		let defaultConfig = {
+    super(props, {
 
       idea:{
         showVoteButtons: true, // TODO: dit is een stomme config parameter
@@ -24,6 +23,18 @@ export default class IdeasOverview extends OpenStadComponent {
         minimumYesVotes: undefined,
       },
 
+      display: {
+        type: 'tiles',
+        columns: 4,
+        showStatusLabel: false,
+        showTheme: false,
+        showArea: false,
+        showVoteProgressbar: false,
+        showStats: false,
+        showTypeIcon: false,
+        onMouseOverTileFadeOthers: false,
+      },
+      
       image: {},
 
       argument: {
@@ -39,46 +50,7 @@ export default class IdeasOverview extends OpenStadComponent {
       types: [],
       typeField: 'typeId',
 
-		};
-
-    // type specific default config
-    switch (props.config.display && props.config.display.type) {
-      case 'list':
-        defaultConfig.display = {
-          columns: 1,
-          showStatusLabel: false,
-          showTheme: false,
-          showArea: false,
-          showVoteProgressbar: false,
-          showStats: true,
-          showTypeIcon: true,
-        };
-        break;
-      case 'grid':
-        defaultConfig.display = {
-          columns: 3,
-          showStatusLabel: false,
-          showTheme: true,
-          showArea: true,
-          showVoteProgressbar: false,
-          showStats: false,
-          showTypeIcon: false,
-        };
-        break;
-      default:
-        defaultConfig.display = {
-          type: 'tiles',
-          columns: 4,
-          showStatusLabel: true,
-          showTheme: false,
-          showArea: false,
-          showVoteProgressbar: true,
-          showStats: true,
-          showTypeIcon: false,
-        };
-    }
-
-    super(props, defaultConfig);
+		});
 
     this.config.image.allowMultipleImages = false;
 
@@ -200,7 +172,7 @@ export default class IdeasOverview extends OpenStadComponent {
 
     // now render
     return (
-			<div id={self.divId} className={`osc-idea-tile osc-displaytype-${self.config.display.type} ${self.props.className || ''}`} onClick={event => self.dispatchIdeaTileClick(event, idea)} onMouseOver={event => self.dispatchIdeaTileMouseOver(event, idea)} onMouseOut={event => self.dispatchIdeaTileMouseOut(event, idea)}>
+			<div id={self.divId} className={`osc-idea-tile osc-displaytype-${self.config.display.type} osc-column osc-${self.config.display.columns}-columns ${self.props.className || ''}`} onClick={event => self.dispatchIdeaTileClick(event, idea)} onMouseOver={event => self.dispatchIdeaTileMouseOver(event, idea)} onMouseOut={event => self.dispatchIdeaTileMouseOut(event, idea)}>
         <div className="osc-idea-image-container">
           <OpenStadComponentIdeaImage config={this.config} idea={idea} key={'image-' + idea.id}/>
         </div>