+++ src/choices-guide/component/question-group.jsx
@@ -64,7 +64,7 @@ export default class OpenStadComponentQuestionGroup extends OpenStadComponent {
 
   gotoNextQuestion() {
 
-    if (!this.isValid()) return;
+    if (!this.isValid()) return {};
 
     let state = {};
     let currentQuestion = this.state.currentQuestion + this.noOfQuestionsToShow;